% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHAgricProductivityLab_control.R
\name{GHAgricProductivityLab_control}
\alias{GHAgricProductivityLab_control}
\title{Control Settings (Internal)}
\usage{
GHAgricProductivityLab_control(number_of_draws = 100, myseed = 11122025)
}
\arguments{
\item{number_of_draws}{Integer(1). Number of draws.}

\item{myseed}{Numeric(1). Random seed.}
}
\value{
A list containing control parameters.
}
\description{
Creates a control list for simulation settings. Used internally.
}
\keyword{internal}
