% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mode}
\alias{mode}
\title{Mode of a Vector (Internal)}
\usage{
mode(x, na.rm = TRUE)
}
\arguments{
\item{x}{Vector.}

\item{na.rm}{Logical; remove NA values. Default TRUE.}
}
\value{
The modal value.
}
\description{
Computes the most frequent value in a vector. Used internally.
}
\keyword{internal}
