% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okwaayeli-package.R
\docType{package}
\name{okwaayeli-package}
\alias{okwaayeli}
\alias{okwaayeli-package}
\title{okwaayeli: Agricultural Productivity in Ghana}
\description{
Provides tools and datasets for investigating the drivers of agricultural production shortfalls in Ghana. It compiles research studies that examine farmer-specific and institutional factors to assess whether inefficiencies arise from technical inefficiency, technology gaps, or both. The package offers empirical evidence to inform policy discussions and interventions aimed at improving agricultural productivity, particularly in contexts with limited access to modern technologies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ftsiboe/okwaayeli}
  \item Report bugs at \url{https://github.com/ftsiboe/okwaayeli/issues}
}

}
\keyword{internal}
